@@ -19,40 +19,40 @@
   "types": "lib/cjs/common/index.d.ts",
   "private": true,
   "dependencies": {
-    "@bugsnag/js": "^7.10.5",
+    "@bugsnag/js": "^7.11.0",
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
-    "jwks-rsa": "^2.0.3",
-    "mongodb": "^3.3.5",
+    "jwks-rsa": "^2.0.4",
+    "mongodb": "^4.1.3",
     "ot-json0": "^1.1.0",
     "rich-text": "^3.1.0",
     "sharedb": "1.0.0-beta.31",
     "sharedb-access": "^5.0.0",
-    "sharedb-mongo": "^1.0.0-beta.17",
+    "sharedb-mongo": "1.0.0-beta.21",
     "ts-object-path": "^0.1.2",
     "websocket-json-stream": "^0.0.3",
-    "ws": "^7.4.6"
+    "ws": "^8.2.3"
   },
   "devDependencies": {
-    "@types/jest": "^26.0.23",
+    "@types/jest": "^27.0.2",
     "@types/jest-expect-message": "^1.0.3",
-    "@types/jsonwebtoken": "^8.5.4",
-    "@types/mongodb": "^3.3.11",
-    "@types/ws": "^7.4.5",
-    "@typescript-eslint/eslint-plugin": "^4.29.3",
-    "@typescript-eslint/parser": "^4.29.3",
+    "@types/jsonwebtoken": "^8.5.5",
+    "@types/ws": "^8.2.0",
+    "@typescript-eslint/eslint-plugin": "^4.33.0",
+    "@typescript-eslint/parser": "^4.33.0",
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "2.22.1",
-    "eslint-plugin-jsdoc": "30.7.6",
-    "eslint-plugin-prettier": "^3.4.1",
-    "jest": "^27.0.4",
+    "eslint-plugin-deprecation": "^1.2.1",
+    "eslint-plugin-import": "^2.24.2",
+    "eslint-plugin-jsdoc": "^36.1.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "jest": "^27.2.4",
     "jest-expect-message": "^1.0.2",
     "jest-teamcity-reporter": "^0.9.0",
-    "prettier": "^2.3.2",
-    "sharedb-mingo-memory": "^1.1.4",
-    "ts-jest": "^27.0.3",
+    "prettier": "^2.4.1",
+    "sharedb-mingo-memory": "^1.2.0",
+    "ts-jest": "^27.0.5",
     "ts-mockito": "^2.6.1",
-    "typescript": "~4.1.5"
+    "typescript": "~4.3.5"
   }
 }
