@@ -14,45 +14,45 @@
     "lint": "ng lint",
     "prettier": "prettier --write \"**/*.{ts,js,json,css,scss,less,html,md,yml}\"",
     "prettier:tc": "prettier --list-different \"**/*.{ts,js,json,css,scss,less,html,md,yml}\"",
-    "e2e": "ng e2e",
     "bundle-report": "webpack-bundle-analyzer dist/stats.json",
-    "build:stats": "ng build --prod --stats-json"
+    "build:stats": "ng build --configuration production --stats-json",
+    "prepare": "cd ../../.. && husky install src/SIL.XForge.Scripture/ClientApp/.husky"
   },
   "private": true,
   "dependencies": {
     "@angular-mdc/web": "^5.1.1",
-    "@angular/animations": "~11.2.14",
-    "@angular/cdk": "~11.2.13",
-    "@angular/common": "~11.2.14",
-    "@angular/core": "~11.2.14",
-    "@angular/flex-layout": "~11.0.0-beta.33",
-    "@angular/forms": "~11.2.14",
-    "@angular/material": "~11.2.13",
-    "@angular/platform-browser": "~11.2.14",
-    "@angular/platform-browser-dynamic": "~11.2.14",
-    "@angular/pwa": "~0.1102.18",
-    "@angular/router": "~11.2.14",
-    "@angular/service-worker": "~11.2.14",
+    "@angular/animations": "~12.2.16",
+    "@angular/cdk": "~12.2.13",
+    "@angular/common": "~12.2.16",
+    "@angular/core": "~12.2.16",
+    "@angular/flex-layout": "~12.0.0-beta.35",
+    "@angular/forms": "~12.2.16",
+    "@angular/material": "~12.2.13",
+    "@angular/platform-browser": "~12.2.16",
+    "@angular/platform-browser-dynamic": "~12.2.16",
+    "@angular/pwa": "~12.2.16",
+    "@angular/router": "~12.2.16",
+    "@angular/service-worker": "~12.2.16",
     "@bugsnag/js": "^7.16.1",
     "@bugsnag/plugin-angular": "^7.16.1",
-    "@ngneat/transloco": "^2.23.5",
-    "@ngneat/transloco-locale": "^1.4.0",
-    "@sillsdev/machine": "^2.2.2",
-    "angular-file": "^3.4.0",
-    "angular-split": "^4.0.0",
+    "@ngneat/transloco": "^3.0.2",
+    "@ngneat/transloco-locale": "^3.0.1",
+    "@sillsdev/machine": "^2.2.3",
+    "angular-file": "^3.6.0",
+    "angular-split": "^5.0.0",
     "arraydiff": "^0.1.3",
     "auth0-js": "^9.19.0",
     "bootstrap": "^4.6.1",
     "bowser": "^2.11.0",
-    "bson-objectid": "^1.3.1",
+    "bson-objectid": "^2.0.1",
     "buffer": "^5.7.1",
     "crc-32": "^1.2.1",
     "jwt-decode": "^3.1.2",
     "lodash-es": "^4.17.21",
-    "mingo": "^2.5.3",
+    "mingo": "^4.2.0",
     "ngx-avatar": "^4.1.0",
-    "ngx-cookie-service": "^2.4.0",
-    "ngx-quill": "^13.4.0",
+    "ngx-cookie-service": "~12.0.3",
+    "ngx-quill": "14.1.1",
     "ot-json0": "^1.1.0",
     "papaparse": "^5.3.1",
     "process": "^0.11.10",
@@ -88,37 +89,31 @@
     "@types/node": "^12.20.46",
     "@types/papaparse": "^5.3.2",
     "@types/quill": "^1.3.10",
-    "@types/uuid": "3.4.6",
-    "@typescript-eslint/eslint-plugin": "4.16.1",
-    "@typescript-eslint/parser": "4.16.1",
+    "@types/uuid": "8.3.1",
+    "@typescript-eslint/eslint-plugin": "^4.33.0",
+    "@typescript-eslint/parser": "^4.33.0",
     "codecov": "^3.8.3",
-    "eslint": "^7.6.0",
+    "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "2.22.1",
-    "eslint-plugin-jsdoc": "30.7.6",
-    "eslint-plugin-prettier": "^3.4.1",
-    "husky": "^4.3.8",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.2",
+    "eslint-plugin-deprecation": "^1.2.1",
+    "eslint-plugin-import": "^2.24.2",
+    "eslint-plugin-jsdoc": "^36.1.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "husky": "^7.0.2",
+    "jasmine-core": "~3.9.0",
     "karma": "~6.3.16",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "^3.0.3",
     "karma-jasmine": "~4.0.1",
-    "karma-jasmine-html-reporter": "~1.5.4",
+    "karma-jasmine-html-reporter": "~1.7.0",
     "karma-teamcity-reporter": "^1.1.0",
-    "lint-staged": "^10.5.4",
-    "npm": "^7.24.2",
-    "prettier": "^2.3.2",
-    "protractor": "~7.0.0",
+    "lint-staged": "^11.2.0",
+    "npm": "^8.1.0",
+    "prettier": "^2.4.1",
     "ts-mockito": "^2.6.1",
-    "ts-node": "~8.8.2",
-    "typescript": "~4.1.6",
+    "ts-node": "^10.2.1",
+    "typescript": "~4.3.5",
     "webpack-bundle-analyzer": "^4.5.0"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
   },
   "lint-staged": {
     "*.{ts,js,json,css,scss,less,html,md,yml}": [
