@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { timer } from 'rxjs';
-import { map, takeWhile } from 'rxjs/operators';
+import { takeWhile } from 'rxjs/operators';
 
 @Component({
   selector: 'app-page-not-found',
@@ -11,8 +11,7 @@ import { map, takeWhile } from 'rxjs/operators';
 export class PageNotFoundComponent {
   // Timer fires every 100ms until 100 instances have occurred (10s)
   progress = timer(0, 100).pipe(
-    map(count => count / 100),
-    takeWhile(val => (val <= 1 ? true : this.router.navigateByUrl('/projects') && false))
+    takeWhile(val => (val <= 100 ? true : this.router.navigateByUrl('/projects') && false))
   );
 
   constructor(readonly router: Router) {}
