@@ -626,9 +626,7 @@ class TestEnvironment {
   }
 
   openFromScriptureChooser(): void {
-    this.click(
-      this.overlayContainerElement.querySelector('mdc-text-field[formControlName="from"] mdc-icon') as HTMLInputElement
-    );
+    this.click(this.overlayContainerElement.querySelector('#from-btn') as HTMLInputElement);
   }
 
   setOnline(value: boolean) {
@@ -684,7 +682,7 @@ class TestEnvironment {
 
   private simulateTransceleratorQuestionsAlreadyExisting(ids: number[]): void {
     ids.forEach(id => {
-      const doc: TransceleratorQuestion = this.questions.find(question => question.id == id + '')!;
+      const doc: TransceleratorQuestion = this.questions.find(question => question.id === id + '')!;
       const verse = doc.endVerse && !doc.endChapter ? doc.startVerse + '-' + doc.endVerse : doc.startVerse;
       this.existingQuestions.push({
         data: {
