@@ -2,7 +2,7 @@ import { translate } from '@ngneat/transloco';
 import Bowser from 'bowser';
 import ObjectID from 'bson-objectid';
 import locales from '../../../locales.json';
-import { version } from '../../../version.json';
+import versionData from '../../../version.json';
 import { environment } from '../environments/environment';
 import { Locale } from './models/i18n-locale';
 
@@ -12,8 +12,11 @@ export function nameof<T>(name: Extract<keyof T, string>): string {
   return name;
 }
 
+/**
+ * Returns a valid 24 character `ObjectID` hex string.
+ */
 export function objectId(): string {
-  return ObjectID.generate();
+  return new ObjectID().toHexString();
 }
 
 export function supportedBrowser(): boolean {
@@ -48,10 +51,10 @@ export function getBrowserEngine(): string {
 
 export function issuesEmailTemplate(errorId?: string): string {
   const bowser = Bowser.getParser(window.navigator.userAgent);
-  const subject = translate('issue_email.subject', { siteName: environment.siteName });
-  const body = translate('issue_email.body', {
+  const subject: string = translate('issue_email.subject', { siteName: environment.siteName });
+  const body: string = translate('issue_email.body', {
     siteName: environment.siteName,
-    siteVersion: version,
+    siteVersion: versionData.version,
     browserName: bowser.getBrowserName(),
     browserVersion: bowser.getBrowserVersion(),
     operatingSystem: bowser.getOSName(),
